Matching size for putSram
putSram requires specific sizes for each entry

RTC: 254766
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/100711
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/100734
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>

+++ src/import/chips/p10/procedures/hwp/nest/p10_io_load_ppe.C
@@ -140,7 +140,7 @@ fapi2::ReturnCode p10_io_load_ppe(
                          PAU0_PERV_CHIPLET_ID,  // Set to PAU0 as default for multicast.
                          // p10_putsram will find valid multicast target.
                          true,                  // Do multicast load
-                         0,                     // Access mode, not used here
+                         static_cast<uint8_t>(0), // Access mode, not used here
                          SRAM_IO_PPE_IMAGE_OFFSET,
                          l_ioImgSize,
                          l_ioImgDataPtr);
@@ -157,7 +157,7 @@ fapi2::ReturnCode p10_io_load_ppe(
                              PAU0_PERV_CHIPLET_ID, // Set to PAU0 as default for multicast.
                              // p10_putsram will find valid multicast target.
                              true,                 // Do multicast load
-                             0,                    // Access mode, not used here
+                             static_cast<uint8_t>(0), // Access mode, not used here
                              MEM_REG_OFFSETS[ii],
                              l_memRegsImgSize,
                              l_memRegsImgDataPtr);
